Move pkg/gitutils to remotecontext/git
