+++ builder/remotecontext/git.go
@@ -4,13 +4,13 @@ import (
 	"os"
 
 	"github.com/docker/docker/builder"
+	"github.com/docker/docker/builder/remotecontext/git"
 	"github.com/docker/docker/pkg/archive"
-	"github.com/docker/docker/pkg/gitutils"
 )
 
 // MakeGitContext returns a Context from gitURL that is cloned in a temporary directory.
 func MakeGitContext(gitURL string) (builder.Source, error) {
-	root, err := gitutils.Clone(gitURL)
+	root, err := git.Clone(gitURL)
 	if err != nil {
 		return nil, err
 	}
